refactor: shift copyBytes and tweak deps to reduce dependencies
